fix(howl): parse HOWL/CSEQ headers at host pointer width

+++ game/HOWL/HOWL_Load.c
@@ -58,7 +58,7 @@ int howl_InitGlobals(char *filename)
 // NOTE(aalhendi): ASM-verified NTSC-U 926 0x80029a50-0x80029ab4
 void howl_ParseHeader(struct HowlHeader *hh)
 {
-	u32 addr = (u32)hh;
+	uintptr_t addr = (uintptr_t)hh;
 
 	sdata->ptrHowlHeader = (struct HowlHeader *)addr;
 	addr += sizeof(struct HowlHeader);
@@ -78,13 +78,13 @@ void howl_ParseHeader(struct HowlHeader *hh)
 	sdata->howl_songOffsets = (u16 *)addr;
 	addr += sizeof(s16) * hh->numSequences;
 
-	sdata->howl_endOfHowl = addr;
+	sdata->howl_endOfHowl = (int)addr;
 }
 
 // NOTE(aalhendi): ASM-verified NTSC-U 926 0x80029ab4-0x80029b2c
 void howl_ParseCseqHeader(struct CseqHeader *ch)
 {
-	u32 addr = (u32)ch;
+	uintptr_t addr = (uintptr_t)ch;
 
 	sdata->ptrCseqHeader = (struct CseqHeader *)addr;
 	addr += sizeof(struct CseqHeader);
@@ -137,7 +137,7 @@ int howl_LoadHeader(char *filename)
 			MEMPACK_ReallocMem(numSector << 0xb);
 
 			// if header needs more sectors loaded, like CTR-U which needs 3 sectors
-			if (numSector < 2 || LOAD_HowlHeaderSectors(&sdata->KartHWL_CdFile, (void *)((int)alloc + 0x800), 1, numSector - 1) != 0)
+			if (numSector < 2 || LOAD_HowlHeaderSectors(&sdata->KartHWL_CdFile, (void *)((uintptr_t)alloc + 0x800), 1, numSector - 1) != 0)
 			{
 				// initilaize header and pointer table
 				howl_ParseHeader(alloc);
